helpppp

+++ src/main/java/ru/itis/sockets/dao/CrudDao.java
@@ -4,10 +4,10 @@ import java.util.List;
 import java.util.Optional;
 
 public interface CrudDao<T> {
-    Optional<T> find(Integer id);
+    Optional<T> find(Long id);
     void save(T model);
     void update(T model);
-    void delete(Integer id);
+    void delete(Long id);
 
     List<T> findAll();
 }
